상호작용 오브젝트 추가, 문의 상호작용 완성
맵 파일을 불러올 때 objectInfo에 따라 상호작용이 가능한 오브젝트들은 Zone에서 따로 포인터를 가지고 있고, 움직일 수 있는 오브젝트들 이므로 인스턴싱 대신 직접 렌더하도록 바꾸었다. 플레이어는 바라보는 방향에 있는 상호작용 가능한 오브젝트들 중 가장 가까운 것과 상호작용을 할수 있도록 하였다. 문이 열릴 때의 밀림도 잘 된다. 모든 상호작용 오브젝트는 서버에게 질의를 한 후 답을 받았을 때 실제로 작동할 것 이다.

+++ AvoidHimClient/Player.cpp
@@ -41,7 +41,7 @@ void Player::Render(const ComPtr<ID3D12GraphicsCommandList>& _pCommandList) {
 	GameObject::Render(_pCommandList);
 }
 
-void Player::Animate(char _collideCheck, double _timeElapsed) {
+void Player::Animate(char _collideCheck, float _timeElapsed) {
 
 	sendMovePacketTime += (float)_timeElapsed;
 	// y방향으로 충돌하지 않을 경우
@@ -102,7 +102,7 @@ shared_ptr<Camera> Player::GetCamera() {
 }
 
 
-void Player::UpdateRigidBody(double _timeElapsed) {
+void Player::UpdateRigidBody(float _timeElapsed) {
 
 	if (!landed) {
 		velocity.y -= GRAVITY * _timeElapsed;
